Few Gen 5 sprites and refined sprite-bounce
The bounce process is a bit simpler due to a "bounce" property being added to ANIMCMD_FRAME. The process now just checks to see if the bounce property is set to TRUE, if the y2 value of the sprite is not already being changed, and if an affine animation is not currently underway. If all of that's true, it will apply the bounce.